Add'l Sidebar Info.
Filled out the sidebar. The on-click seems broken, but I was trying to figure out what's some of these other options weren't as clear are. @mikelmaron Are you able to troubleshoot? I'm a bit stumped.

Aslo open to suggestions of what else to add into this sidebar that I'm missing.